+++ src/torchwrench/__main__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from .entries import print_install_info
+from .entrypoints.info import main_info
 
 if __name__ == "__main__":
-    print_install_info()
+    main_info()