+++ src/torchwrench/serialization/torch.py
@@ -10,12 +10,9 @@ from pathlib import Path
 from typing import IO, Any, BinaryIO, Callable, Dict, Optional, Union
 
 import torch
-from pythonwrench._core import _setup_output_fpath
 from pythonwrench.semver import Version
-from torch import (  # noqa: F401
-    load,
-    save,
-)
+from pythonwrench.serialization._core import _setup_output_fpath
+from torch import load, save  # noqa: F401
 from torch.serialization import DEFAULT_PROTOCOL
 from torch.types import Storage
 from typing_extensions import TypeAlias