+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.9.0] 2026-07-24
+### Added
+- `set_seed` function.
+
+### Modified
+- `as_dtype` and `as_device` now supports 'none' and 'null' strings.
+- Upgrade pythonwrench dependancy to `>=0.6.0,<0.7.0`.
+
 ## [0.8.0] 2026-03-05
 ### Modified
 - Lazily load imports with `lazy-loader` package.